Enterprise Multimodal QA & OCR/requirements.txt
# Core Framework
streamlit>=1.28.0

# RAG & Vector Store
chromadb>=0.4.22
sentence-transformers>=2.3.1

# Document Processing
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0

# API & Requests
requests>=2.31.0
python-dotenv>=1.0.0

# Data Processing
numpy>=1.24.0

# ML Dependencies (required by sentence-transformers)
torch>=2.1.0
transformers>=4.36.0